--- a/.codespellrc
+++ b/.codespellrc
@@ -1,4 +1,4 @@
 # See: https://github.com/codespell-project/codespell#using-a-config-file
 [codespell]
-skip = ./node_modules,./build,yarn.lock
+skip = ./node_modules,./build,pnpm-lock.yaml
 contex=6