--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -16,7 +16,7 @@ jobs:
       - name: Setup Node 
         uses: actions/setup-node@v4.0.2
         with:
-          node-version: 16
+          node-version: 18
 
       - run: yarn install