--- a/README.md
+++ b/README.md
@@ -2,15 +2,15 @@ This is the source for the SerenityOS Discord Bot.
 
 ### Setup
 
-The bot is written in [TypeScript](https://www.typescriptlang.org), `nodejs` and `yarn` are pre-requisites.
+The bot is written in [TypeScript](https://www.typescriptlang.org), `nodejs` and `pnpm` are pre-requisites.
 
 Then setup your environment:
 
 ```
 $ git clone https://github.com/SerenityOS/discord-bot
 $ cd discord-bot
-$ yarn install
-$ yarn build
+$ pnpm install
+$ pnpm build
 ```
 
 ### Configuration
@@ -23,16 +23,7 @@ guild_id=<your-guild-id-goes-here>
 ```
 See: https://www.writebots.com/discord-bot-token/
 
-Now you can run `yarn start:dev` and the bot will startup, and then restart as you save changes to the source files:
-```
-$ yarn start:dev
-[nodemon] 2.0.7
-[nodemon] to restart at any time, enter `rs`
-[nodemon] watching path(s): src/**/*
-[nodemon] watching extensions: ts,js
-[nodemon] starting `ts-node ./src/index.ts`
-Buggie bot has started
-```
+Now you can run `pnpm start:dev` and the bot will startup, and then restart as you save changes to the source files.
 
 ### Running Tests