Bringing a lot of things back up to the latest and greatest in JS land

Performance differences (measured on a rather old Thinkpad T430, so your times will probably be faster):
- Package install time (`(yarn,pnpm) install`): 17.46s -> 2.4s
- Build time (`build` script): 6.9s -> 11ms
- Test time: (`test` script): 6.34s -> 1.56s