@CommandMC 2024-05-18
- `.pre-commit-config.yml` is updated: <https://github.com/CommandMC/serenity-discord-bot/blob/9c881c3897a22d9ac45225d1f91b075b25e47c64/.pre-commit-config.yaml#L30>.
- `npm` is generally fine to use, as long as you don't use it to manage packages (e.g. running scripts like `start` or `lint` is fine). In this case, I don't think Heroku preinstalls `pnpm`, so I just left that file as-is.
- Yes, no changes are necessary there ([GH Docs on supported package managers](<https://docs.github.com/en/enterprise-cloud@latest/code-security/dependabot/dependabot-version-updates/about-dependabot-version-updates#supported-repositories-and-ecosystems>)). I'm actually not sure whether the "Supported versions" list there is accurate, I've been using pnpm 9 on another project and am still getting dependabot updates.